[tmp/review] Extract SSL test into end-to-end tests (run only in CI)
