+++ flink-end-to-end-tests/pom.xml
@@ -78,6 +78,7 @@ under the License.
 		<module>flink-end-to-end-tests-sql</module>
 		<module>flink-end-to-end-tests-jdbc-driver</module>
 		<module>flink-end-to-end-tests-hive</module>
+		<module>flink-end-to-end-tests-restclient</module>
     </modules>
 
 	<dependencies>
